requirements.txt
django
groq
google-genai
